Correção arquitetural da função de exclusão

+++ src/useCases/DeleteACCType/DeleteACCTypeUseCase.ts
@@ -9,6 +9,11 @@ export class DeleteACCTypeUseCase {
   }
 
   async execute(data: IDeleteACCTypeRequestDTO): Promise<void> {
+    const accsLength = await this.accTypesRepository.getACCsLength(data);
+
+    if (accsLength > 0)
+      throw new Error('Este Tipo de ACC possui ACCs associadas a ele.');
+
     await this.accTypesRepository.delete(data);
   }
 }
